Have most maintence only run in production.

+++ app/Console/Commands/DailyMaintenance.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Commands;
 
+use App\Jobs\PurgeOrphanImages;
 use App\Jobs\UpdateRebrickable;
 use App\LDraw\PartManager;
 use App\Models\Part\Part;
@@ -30,64 +31,45 @@ class DailyMaintenance extends Command
      */
     public function handle()
     {
-        $this->info('Reloading all subparts');
-        Part::lazy()->each(fn (Part $p) => app(PartManager::class)->loadSubparts($p));
+        // None of these need to happen daily on the dev server
+        // and just soaks up resources.
+        if (app()->environment('production')) {
+            $this->info('Reloading all subparts');
+            Part::lazy()->each(fn (Part $p) => app(PartManager::class)->loadSubparts($p));
 
-        $this->info('Recounting all votes');
-        Part::unofficial()->lazy()->each(fn (Part $p) => $p->updatePartStatus());
+            $this->info('Recounting all votes');
+            Part::unofficial()->lazy()->each(fn (Part $p) => $p->updatePartStatus());
 
-        $this->info('Rechecking all unofficial parts');
-        $this->call('lib:check', ['--unofficial-only' => true]);
+            $this->info('Rechecking all unofficial parts');
+            $this->call('lib:check', ['--unofficial-only' => true]);
 
-        $this->info('Removing orphan keywords');
-        PartKeyword::doesntHave('parts')->delete();
+            $this->info('Removing orphan keywords');
+            PartKeyword::doesntHave('parts')->delete();
 
-        $this->info('Removing orphan images');
-        $images = Storage::disk('images')->allFiles('library/unofficial');
-        $files = collect($images)
-            ->map(function (string $file): string {
-                $file = str_replace('_thumb.png', '.png', $file);
-                if (strpos($file, 'textures/') !== false) {
-                    return str_replace('library/unofficial/', '', $file);
-                } else {
-                    return str_replace(['library/unofficial/', '.png'], ['', '.dat'], $file);
-                }
-            })
-            ->unique()
-            ->all();
-        $in_use_files = Part::unofficial()
-            ->whereIn('filename', $files)
-            ->pluck('filename')
-            ->map(
-                fn (string $filename): string =>
-                    strpos($filename, 'textures/') !== false ? "library/unofficial/{$filename}" : str_replace('.dat', '.png', "library/unofficial/{$filename}")
-            );
-        foreach ($images as $image) {
-            if (!$in_use_files->contains($image) && !$in_use_files->contains(str_replace('_thumb.png', '.png', $image))) {
-                Storage::disk('images')->delete($image);
-            }
+            $this->info('Refreshing Rebrickable data');
+            Part::unofficial()
+                ->canHaveExternalData()
+                ->where(fn ($query) =>
+                    $query
+                        ->orwhereBetween('created_at', [now(), now()->subDay()])
+                        ->orWhereBetween('created_at', [now()->subWeek()->subDay(), now()->subWeek()])
+                        ->orWhereBetween('created_at', [now()->subMonth()->subDay(), now()->subMonth()])
+                )
+                ->each(fn (Part $part) => UpdateRebrickable::dispatch($part));
+
+            $this->info('Reloading colors for LDConfig');
+            $this->call('lib:update-colours');
+
+            $this->info('Regenerate unofficial zip');
+            $this->call('lib:refresh-zip');
         }
 
-        $this->info('Refreshing Rebrickable data');
-        Part::unofficial()
-            ->canHaveExternalData()
-            ->where(fn ($query) =>
-                $query
-                    ->orwhereBetween('created_at', [now(), now()->subDay()])
-                    ->orWhereBetween('created_at', [now()->subWeek()->subDay(), now()->subWeek()])
-                    ->orWhereBetween('created_at', [now()->subMonth()->subDay(), now()->subMonth()])
-            )
-            ->each(fn (Part $part) => UpdateRebrickable::dispatch($part));
+        $this->info('Removing orphan images');
+        PurgeOrphanImages::dispatch();
 
         $this->info('Queueing missing images');
         $this->call('lib:render-parts', ['--unofficial-only' => true, '--missing' => true]);
         $this->call('lib:render-models', ['--missing' => true]);
-
-        $this->info('Reloading colors for LDConfig');
-        $this->call('lib:update-colours');
-
-        $this->info('Regenerate unofficial zip');
-        $this->call('lib:refresh-zip');
 
         $this->info('Pruning failed jobs');
         $this->call('queue:prune-failed');
